requirements.txt
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.26.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
matplotlib==3.10.3
mplcyberpunk==0.7.6
newsapi-python==0.2.7
newsdataapi==0.1.25
openai==1.97.0
requests==2.32.4
yfinance==0.2.65
python-dotenv==1.0.0
